ivp_utility: x86_64 for strlen
Change p_strlen return type from int to ptrdiff_t to avoid truncation
on 64-bit platforms. Also change ch from int to char in GBS_remove_escape.

Based on: Source-Authors/source-physics@3233384

+++ ivp_utility/ivu_string.hxx
@@ -30,7 +30,7 @@ extern char *p_make_string(const char *templat, ...);
 extern char *p_make_string_fast(const char *templat, ...); // doesn't check overflow
 
 int p_strcmp(const char *s1, const char *s2);
-int p_strlen(const char *s);
+ptrdiff_t p_strlen(const char *s);
 char *p_read_first_token(FILE *fp);
 char *p_get_next_token();
 int p_get_num();
